requirements.txt
Pillow==9.5.0
Pillow==10.0.0
Pillow==10.1.0
streamlit==1.27.1
streamlit_javascript==0.1.5
streamlit_option_menu==0.3.6
